pybricks.common.BLE: add broadcast/observe methods
This adds new methods to broadcast and observe advertising data using
Bluetooth Low Energy. The communication protocol uses the LEGO
manufacturer-specific data similar to the experimental protocol
from the official Robot Inventor firmware. Multiple Python objects
are serialized using a compressed format to squeeze as many objects
into the 31 byte advertising data as possible.

Since the BLE object is a singleton initialized on the hub object, the
initialization parameters are added to the Hub constructors.

Each hub can only broadcast on a single "channel" and can receive on
multiple channels. Channels are limited to 16 to avoid allocating too
much memory.